+++ src/components/dialog/dialog.tsx
@@ -1,6 +1,7 @@
 import { Component, Element, Event, EventEmitter, Method, Prop, State, Watch, h } from '@stencil/core';
 import { lockBodyScrolling, unlockBodyScrolling } from '../../utilities/scroll';
 import { hasSlot } from '../../utilities/slot';
+import Modal from '../../utilities/modal';
 
 let id = 0;
 
@@ -29,6 +30,7 @@ let id = 0;
 export class Dialog {
   componentId = `dialog-${++id}`;
   dialog: HTMLElement;
+  modal: Modal;
   panel: HTMLElement;
 
   @Element() host: HTMLSlDialogElement;
@@ -72,12 +74,15 @@ export class Dialog {
   @Event({ eventName: 'sl-overlay-dismiss' }) slOverlayDismiss: EventEmitter;
 
   connectedCallback() {
-    this.handleDocumentFocusIn = this.handleDocumentFocusIn.bind(this);
     this.handleCloseClick = this.handleCloseClick.bind(this);
     this.handleTransitionEnd = this.handleTransitionEnd.bind(this);
     this.handleKeyDown = this.handleKeyDown.bind(this);
     this.handleOverlayClick = this.handleOverlayClick.bind(this);
     this.updateSlots = this.updateSlots.bind(this);
+
+    this.modal = new Modal(this.host, {
+      onFocusOut: () => this.panel.focus()
+    });
   }
 
   componentWillLoad() {
@@ -114,9 +119,9 @@ export class Dialog {
 
     this.isVisible = true;
     this.open = true;
+    this.modal.activate();
 
     lockBodyScrolling(this.host);
-    document.addEventListener('focusin', this.handleDocumentFocusIn);
   }
 
   /** Hides the dialog */
@@ -134,21 +139,13 @@ export class Dialog {
     }
 
     this.open = false;
+    this.modal.deactivate();
 
     unlockBodyScrolling(this.host);
-    document.removeEventListener('focusin', this.handleDocumentFocusIn);
   }
 
   handleCloseClick() {
     this.hide();
-  }
-
-  handleDocumentFocusIn(event: Event) {
-    const target = event.target as HTMLElement;
-
-    if (target.closest('sl-dialog') !== this.host) {
-      this.panel.focus();
-    }
   }
 
   handleKeyDown(event: KeyboardEvent) {